Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    long a;
-    scanf("%ld",&a);
-    printf("You entered: %ld", a);
+    double a;
+    scanf("%lf",&a);
+    printf("You entered: %.4lf", a);
     return 0;
 }
